LayoutTests/platform/win/http/tests/inspector/resource-har-conversion-expected.txt

Tests conversion of Inspector's resource representation into HAR format.

Page reloaded.
{
    0 : {
        pageref : "http://127.0.0.1:8000/inspector/resource-har-conversion.html"
        startedDateTime : <object>
        time : <number>
        request : {
            method : "GET"
            url : "http://127.0.0.1:8000/inspector/inspector-test2.js"
            headers : <object>
            headersSize : -1
            bodySize : -1
        }
        response : {
            status : 304
            statusText : "Not Modified"
            headers : <object>
            content : {
                size : 0
                mimeType : "application/x-javascript"
            }
            redirectURL : ""
            headersSize : -1
            bodySize : 0
        }
        timings : {
            blocked : -1
            dns : -1
            connect : -1
            send : -1
            wait : <number>
            receive : <number>
            ssl : -1
        }
    }
    1 : {
        pageref : "http://127.0.0.1:8000/inspector/resource-har-conversion.html"
        startedDateTime : <object>
        time : <number>
        request : {
            method : "GET"
            url : "http://127.0.0.1:8000/inspector/resource-har-conversion.html"
            headers : <object>
            headersSize : -1
            bodySize : -1
        }
        response : {
            status : 200
            statusText : "OK"
            headers : <object>
            content : {
                size : 1137
                mimeType : "text/html"
            }
            redirectURL : ""
            headersSize : -1
            bodySize : 1137
        }
        timings : {
            blocked : -1
            dns : -1
            connect : -1
            send : -1
            wait : <number>
            receive : <number>
            ssl : -1
        }
    }
    2 : {
        pageref : "http://127.0.0.1:8000/inspector/resource-har-conversion.html"
        startedDateTime : <object>
        time : <number>
        request : {
            method : "GET"
            url : "http://127.0.0.1:8000/inspector/resource-test2.js"
            headers : <object>
            headersSize : -1
            bodySize : -1
        }
        response : {
            status : 304
            statusText : "Not Modified"
            headers : <object>
            content : {
                size : 0
                mimeType : "application/x-javascript"
            }
            redirectURL : ""
            headersSize : -1
            bodySize : 0
        }
        timings : {
            blocked : -1
            dns : -1
            connect : -1
            send : -1
            wait : <number>
            receive : <number>
            ssl : -1
        }
    }
    3 : {
        pageref : "http://127.0.0.1:8000/inspector/resource-har-conversion.html"
        startedDateTime : <object>
        time : <number>
        request : {
            method : "POST"
            url : "http://127.0.0.1:8000/inspector/resources/xhr-does-not-exist.html"
            headers : <object>
            headersSize : -1
            bodySize : -1
            postData : {
                mimeType : "text/xml"
                text : "<xml></xml>"
            }
        }
        response : {
            status : 404
            statusText : "Not Found"
            headers : <object>
            content : {
                size : 305
                mimeType : "text/html"
            }
            redirectURL : ""
            headersSize : -1
            bodySize : 305
        }
        timings : {
            blocked : -1
            dns : -1
            connect : -1
            send : -1
            wait : <number>
            receive : <number>
            ssl : -1
        }
    }
}
